requirements/base.txt
Django==3.2.7
python-decouple==3.4
djangorestframework==3.12.4